solve: 2562 - 최댓값
